move still relevant parts of headers_tree into amaru-sim
