+++ crates/amaru-consensus/Cargo.toml
@@ -18,20 +18,18 @@ categories.workspace = true
 [features]
 default = ["test-utils"]
 profiling = ["dep:pprof"]
-test-utils = ["proptest", "serde_json", "dep:tracing-subscriber", "dep:pretty_assertions"]
+test-utils = ["proptest", "dep:tracing-subscriber", "dep:pretty_assertions"]
 
 [dependencies]
 async-trait.workspace = true
 anyhow.workspace = true
 hex.workspace = true
-itertools.workspace = true
 opentelemetry.workspace = true
 parking_lot.workspace = true
 pretty_assertions = { workspace = true, optional = true }
 proptest = { workspace = true, optional = true }
 rand = { workspace = true }
 rayon.workspace = true
-serde_json = { workspace = true, optional = true }
 serde.workspace = true
 thiserror.workspace = true
 tokio = { workspace = true, features = ["sync", "rt", "time"] }
@@ -69,11 +67,6 @@ amaru-stores.workspace = true
 
 [target.'cfg(unix)'.dependencies]
 pprof = { version = "0.15", features = ["flamegraph"], optional = true }
-
-[[bench]]
-name = "headers_tree"
-harness = false
-required-features = ["test-utils"]
 
 [[bench]]
 name = "stage_msgs"
