+++ crates/amaru-consensus/src/errors.rs
@@ -33,10 +33,6 @@ pub enum ConsensusError {
     StoreHeaderFailed(HeaderHash, StoreError),
     #[error("Failed to remove header at {0}: {1}")]
     RemoveHeaderFailed(HeaderHash, StoreError),
-    #[error("Failed to set a new anchor at {0}: {1}")]
-    SetAnchorPointFailed(Point, StoreError),
-    #[error("Failed to set a best chain at {0}: {1}")]
-    SetBestChainTipFailed(Point, StoreError),
     #[error("Failed to update a best chain at {0}->{1}: {2}")]
     UpdateBestChainFailed(HeaderHash, HeaderHash, StoreError),
     #[error("Failed to store block body at {0}: {1}")]
