--- crates/amaru-consensus/src/headers_tree/data_generation/mod.rs
@@ -1,23 +1,0 @@
-// Copyright 2025 PRAGMA
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-pub mod actions;
-pub mod data_generation;
-pub mod property_configuration;
-mod shrink;
-
-pub use actions::*;
-pub use data_generation::*;
-pub use property_configuration::*;
-pub use shrink::*;
