--- crates/amaru-consensus/src/headers_tree/headers_tree_display.rs
@@ -1,74 +1,0 @@
-// Copyright 2025 PRAGMA
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use std::fmt::{Debug, Display, Formatter};
-
-use amaru_kernel::{Header, HeaderHash, utils::string::ListToString};
-
-use crate::headers_tree::{HeadersTree, HeadersTreeState, tree::Tree};
-
-/// A displayable version of HeadersTree, for Debug and Display implementations.
-/// It contains a snapshot of both the in memory data for the headers tree but also
-/// its state in the database: anchor and best chain.
-#[derive(Clone)]
-pub struct HeadersTreeDisplay {
-    tree: Option<Tree<Header>>,
-    tree_state: HeadersTreeState,
-    anchor: HeaderHash,
-    best_chain: HeaderHash,
-    best_chains: Vec<HeaderHash>,
-    best_length: usize,
-}
-
-impl From<HeadersTree> for HeadersTreeDisplay {
-    fn from(value: HeadersTree) -> Self {
-        let tree_state = value.clone().into_tree_state();
-        HeadersTreeDisplay {
-            tree: value.to_tree(),
-            tree_state,
-            anchor: value.anchor(),
-            best_chain: value.best_chain(),
-            best_chains: value.best_peers_chains().into_iter().cloned().collect(),
-            best_length: value.best_length(),
-        }
-    }
-}
-
-impl Debug for HeadersTreeDisplay {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        self.format(f, |tip| format!("{tip:?}"))
-    }
-}
-
-impl Display for HeadersTreeDisplay {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        self.format(f, |tip| format!("{tip}"))
-    }
-}
-
-impl HeadersTreeDisplay {
-    /// Common function to either format for Debug or Display
-    pub fn format(&self, f: &mut Formatter<'_>, header_to_string: fn(&Header) -> String) -> std::fmt::Result {
-        f.write_str("HeadersTree {\n")?;
-        if let Some(tree) = &self.tree {
-            writeln!(f, "  headers:\n    {}", tree.pretty_print_with(header_to_string))?;
-        };
-        writeln!(f, "{}", self.tree_state)?;
-        writeln!(f, "  anchor: {}", self.anchor)?;
-        writeln!(f, "  best_chain: {}", self.best_chain)?;
-        writeln!(f, "  best_chains: [{}]", self.best_chains.list_to_string(", "))?;
-        writeln!(f, "  best_length: {}", self.best_length)?;
-        f.write_str("}\n")
-    }
-}
