+++ crates/amaru-consensus/src/lib.rs
@@ -18,7 +18,6 @@ pub mod block_validator;
 pub mod effects;
 pub mod errors;
 pub mod events;
-pub mod headers_tree;
 pub mod performance;
 pub mod span;
 pub mod stages;
