+++ crates/amaru-sim/Cargo.toml
@@ -21,6 +21,7 @@ cbor4ii.workspace = true
 clap.workspace = true
 hex.workspace = true
 parking_lot.workspace = true
+proptest.workspace = true
 rand.workspace = true
 rand_distr.workspace = true
 rayon.workspace = true
@@ -30,7 +31,6 @@ tracing-subscriber.workspace = true
 tracing.workspace = true
 
 # Internal dependencies ───────────────────────────────────────────────────────┐
-amaru-consensus = { workspace = true, features = ["test-utils"] }
 amaru-kernel = { workspace = true, features = ["test-utils"] }
 amaru-node = { workspace = true, features = ["test-utils"] }
 amaru-ouroboros.workspace = true
@@ -42,7 +42,6 @@ amaru-deps.workspace = true
 
 [dev-dependencies]
 # External dependencies ───────────────────────────────────────────────────────┐
-proptest.workspace = true
 rand.workspace = true
 
 [[bin]]
