+++ crates/amaru-sim/src/simulator/data_generation/mod.rs
@@ -1,4 +1,4 @@
-// Copyright 2025 PRAGMA
+// Copyright 2026 PRAGMA
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,8 +12,18 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+pub mod actions;
 pub mod base_generators;
 pub mod generate;
+pub mod headers;
+pub mod property_configuration;
+pub mod shrink;
+pub mod tree;
 
+pub use actions::*;
 pub use base_generators::*;
 pub use generate::*;
+pub use headers::*;
+pub use property_configuration::*;
+pub use shrink::*;
+pub use tree::*;
