+++ crates/amaru-sim/src/simulator/data_generation/property_configuration.rs
