+++ crates/amaru-sim/src/simulator/data_generation/shrink.rs
